add soft exit message when there is no task to clean in utils/clean_task_containers.sh

+++ utils/clean_task_containers.sh
@@ -6,6 +6,12 @@ terminatedTaskIds="$(docker container ls -a \
     --format "{{.ID}}" | \
     sed -z 's/\n/ /g')"
 
+if [[ -z $terminatedTaskIds ]]
+then
+    echo "There is no task container to remove."
+    exit 0
+fi
+
 echo "Removing containers $terminatedTaskIds"
 docker container rm $terminatedTaskIds
 echo "Containers $terminatedTaskIds terminated successfully"
